@@ -2,23 +2,23 @@ import 'package:flutter/rendering.dart';
 
 class RenderSliverGap extends RenderSliver {
   RenderSliverGap({
-    double mainAxisExtent,
-    Color color,
+    double? mainAxisExtent,
+    Color? color,
   })  : _mainAxisExtent = mainAxisExtent,
         _color = color;
 
-  double get mainAxisExtent => _mainAxisExtent;
-  double _mainAxisExtent;
-  set mainAxisExtent(double value) {
+  double? get mainAxisExtent => _mainAxisExtent;
+  double? _mainAxisExtent;
+  set mainAxisExtent(double? value) {
     if (_mainAxisExtent != value) {
       _mainAxisExtent = value;
       markNeedsLayout();
     }
   }
 
-  Color get color => _color;
-  Color _color;
-  set color(Color value) {
+  Color? get color => _color;
+  Color? _color;
+  set color(Color? value) {
     if (_color != value) {
       _color = value;
       markNeedsPaint();
@@ -30,23 +30,23 @@ class RenderSliverGap extends RenderSliver {
     final double paintExtent = calculatePaintOffset(
       constraints,
       from: 0,
-      to: mainAxisExtent,
+      to: mainAxisExtent!,
     );
     final double cacheExtent = calculateCacheOffset(
       constraints,
       from: 0,
-      to: mainAxisExtent,
+      to: mainAxisExtent!,
     );
 
     assert(paintExtent.isFinite);
     assert(paintExtent >= 0.0);
     geometry = SliverGeometry(
-      scrollExtent: mainAxisExtent,
+      scrollExtent: mainAxisExtent!,
       paintExtent: paintExtent,
       cacheExtent: cacheExtent,
-      maxPaintExtent: mainAxisExtent,
+      maxPaintExtent: mainAxisExtent!,
       hitTestExtent: paintExtent,
-      hasVisualOverflow: mainAxisExtent > constraints.remainingPaintExtent ||
+      hasVisualOverflow: mainAxisExtent! > constraints.remainingPaintExtent ||
           constraints.scrollOffset > 0.0,
     );
   }
@@ -54,11 +54,11 @@ class RenderSliverGap extends RenderSliver {
   @override
   void paint(PaintingContext context, Offset offset) {
     if (color != null) {
-      final Paint paint = Paint()..color = color;
+      final Paint paint = Paint()..color = color!;
       final Size size = constraints
           .asBoxConstraints(
-            minExtent: geometry.paintExtent,
-            maxExtent: geometry.paintExtent,
+            minExtent: geometry!.paintExtent,
+            maxExtent: geometry!.paintExtent,
           )
           .constrain(Size.zero);
       context.canvas.drawRect(offset & size, paint);
